IntermidiateProjects/MiniVotingSystem/MiniVotingBackend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Django==5.1.2
djangorestframework==3.15.2
dynaconf==3.2.6
hvac==2.3.0
idna==3.10
knox==0.1.14
loguru==0.7.2
requests==2.32.3
shortuuid==1.0.13
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
validators==0.34.0
